if there's no userAuth (null=false) we wanna return (exit from this function).
If userAuth exists, we wanna query inside of firestore for the document to see if it already exists.

snapShot simply represents the data.

In order to create we use the documentRef object. Any methods of getting, updating or deleting data we have to use the documentRef.

if(!snapShot.exists) = if that user does not exists in our database we are creating it in that place.

And at the end we return that userRef because we could wanna use this user reference to do other things.