A reducer is just a function that gets 2 properties:
- state object (last state/initial state)
- action (object that has a "type", which is a string value, it's just a name that tells us what action this is. And also it has a "payload". payload can be anything)

payload: we might pass an object that we set as the value in our state or we might use this value to make some transformations on our state. It's just a very flexible property that we get on this action object that we can or cannot use.

The state is sth that redux store is gonna pass to this reducer whenever an action fires. State is whatever the state is currently when that action gets fired.