we wanna return the actual state we want based on the action. So we write a switch statement.
We are saying: based on the action.type value, if the case  is what we want, then we will render sth. Otherwhise, by default, we are gonna just return the state.

Every single reducer gets every single action that ever gets fired, even if those actions are not related to this reducer. So the reason we do defautl return state, is because if none of the action types match inside of the switch statement then we just wanna return the state.