The case that we care about is SET_CURRENT_USER.
Whenever SET_CURRENT_USER is the action type that gets fired, we wanna return a new object which represents the new state.

We spread in everything on the state (...state) becasue we only wanna modify the values that we care about and that is currentUser.

The action object is just the type and then a payload. So we are setting the current user value with this payload.

And we export default userReducer.