created store.js file, which will be the state of our app. imported createStore function and applyMiddleware function.
We need to add Middleware to our store so that whenever actions get fired or dispatched we can catch them and then display them. And the middlewares, which are the peace in the middle between our actions firing and our root reducer, are just functions that receive actions in and then do sth with them and then pass them out into the root reducer.