With the new store object that we've instantiated with createStore(), we bring this store into our index.js and pass it into our provider.
The Provider is a component class that we get from react-redux that once passed the store object, we'll be able to give that redux store context to the rest of the app so we can dispatch actions to that store or we can pull values off of the store and into our components and all comes from this Provider.

So we imported store.

We passed store in as a value.

With this code written like this, we now have access to redux inside of our app.