Now we can start implementing the right actions  into our application so that redux can start storing the value of our user into our new userReducer and out of our App.js state.
We have to create the action that is going to trigger the correct case inside of our switch statement.

We are gonna create our action creator functions.

we create user.actions.js.
actions are just functions that return objects. Each object is in the correct format that the action is expected to be.