setCurrentUser function takes 1 parameter: "the user object".
So instead of this.setSate, we are gonna fire off an action that holds that value (user) that we were setting state before.

setCurrentUser function returns an object where:
type: 'SET_CURRENT_USER',

'SET_CURRENT_USER'  is the exact same string that our reducer is expecting. We have to make sure that we always align the action
creators type with the reducer type expectation (case 'SET_CURRENT_USER')  in order to create the appropriate effects in our reducer.