higher order components are just functions that take components as arguments and then return you a new souped up component
So what we do with connect:
connect()(Header)
We pass 2 functions, the 2º one being optional and then that will give us back another higher order component that we pass it our Header.

What we pass as the 1º argument of connect is the function that allows us to access the state with the state being our reducer. our root reducer to be specific.

const mapStateToProps = () => ({})

What we return is an object whre the name of the property is the actual property we wanna pass in and then the value will be the value.

So what we get in thus function is the state object:
const mapStateToProps = state => ({})
This state is the root reducer.

We pass a currentUser property where the value of it will be state
state.user.currentUser:
const mapStateToProps = state => ({
    currentUser: state.user.currentUser
})

We are saying:
- we want the root-reducer (state)
- we want the user value (.user) [root-reducer.js], which will give us our userReducer
- And then we want the currentUser value (user.reducer.js)

And then we pass mapStateToProps to connect:
export default connect(mapStateToProps)(Header);

Now we are getting that null value as current user being passed in as currentUser.