We connect our App to the outcome of our inital connect call using the 2º argument of connect which is mapDispatchToProps.
The 1º argument is mapStateToProps() but our App doesn't need currentUser anymore because outside of passing it into our Header, it only sets it but it doens't do anything with the currentUser in its component itself.

So we pass null as the 1º argument because we don't need any state from our reducer.

The 2º argument is mapDispatchToProps, which is a function that gets this dispatch property and then it will return an object where the prop name will be whatever prop we wanna pass in that dispatches the new action that we're tryna pass, which is setCurrentUser:
const mapDispatchToProps = dispatch => ({})

So we import setCurrentUser because we need that action.

What we'll end up returning inside mapDispatchToProps will be setCurrentUser but it goes to a function that gets the "user" object and then calls dispatch. And "dispacth()" is a way for redux to know that whatever object  you're passing me is gonna be an action object that i'm gonna pass to every reducer:
const mapDispatchToProps = dispatch => ({
  setCurrentUser: user => dispatch()
});

So our user action is a function that gets the user but returns an action object:

//This action sets our current user:
export const setCurrentUser = user => ({
    type: 'SET_CURRENT_USER',
    payload: user
});

So we call our action: dispatch(setCurrentUser(user))
and pass that user in so that we are invokings setCurrentUser with the user that will then be used as the payload:
export const setCurrentUser = user => ({
    type: 'SET_CURRENT_USER',
    payload: user
});

So we are just dispatching the object.

And we pass mapDispatchToProps as the second argument of connect:
export default connect(null, mapDispatchToProps)(App);