And we pass mapStateToProps to connect() so that we have access to this.props.currentUser.
we swittch the signin route to excat.

Instead of : component={SignInAndSignUpPage}
we do: "render"

"render" is like the render() that we use in our class component. It is a JavaScript invocation that determines what component to return in the same place where component would be, instead it'll be javascript.

So we say:

<Route  exact path='/signin' render={() => this.props.currentUser ? (<Redirect to='/'/>) : (<SignInAndSignUpPage/>)} />

based of if there's a currentUser, if true: Redirect to '/', if false: load SignInAndSignUpPage

And now we are not able to access SignIn, it will redirect us to '/' as long as we have already signed in but if we sign out we would be able to go to SignIn and once we SignIn it will take us automatically to '/'.