imported CartDropdown inside header component and placed at the bottom, outside of the div options.
Now we have to conditionally show or hide this CartDropdown so we will trigger sth that will show or hide this component based off of some true or false value

We will created a CartReducer to do this and we will set the value of some property called hidden on that reducer based on clicking this icon.

we wanna keep that in our global state in case anywhere else wants to trigger the ability to hide or show our dropdown.