created cartReducer with case TOGGLE_CART_HIDDEN.
The reason we wanna use toggle vs set is because our cart component will either be in a true or false state, so it is either hidden or not hidden.

hidden: !state.hidden =
whatever this boolean value is, I want you to convert it to the opposite of it. So if it is true, i want it to be false, if it is false, i want it to be true.

And by default we just return the current state.