we're gonna add the item into the array whenever the user clicks it, if there's duplicates they'll just show up in the array as duplicates and we'll handle grouping them later.
We add a new case ADD_ITEM
where we return an object where cartItems is:
cartItems: [] = we want it to be our old cartItems and then with the newest action that got fired, that is triggering our ADD_ITEM, we wanna deposit whatever the item it is in the payload of that action into this array.

First we spread in the existing cartItems:
cartItems: [...state.cartItems]
which are already in our state.

And then we just add the action payload at the end of the array:
cartItems: [...state.cartItems, action.payload]

cartItems: [...state.cartItems, action.payload] =

we are spreading all of our array values and then any additional values we add at the end will automatically appear in the end of this array.

action.payload:
we are getting the item payload, we are assuming that the item coming in will be what will be the payload value on our action.

Now we need to write the action...