Now we need to write the action addItem.
'item' is the item that we wanna addto the array that we made.

type: CartActionTypes.ADD_ITEM, =
this will tell our reducer: I wanna add this item.

payload: item =
and the payload will be the item that we get.

payload can be anything we want it to be and in our case will be the item we are trying to add into our cartItems array.

Now that we have our payload inside of our action definition we will bring this into our components so we can use it...