we need to dispatch this new addItem so we use mapDispatchToProps. We need to take in this new item object.
const mapDispatchToProps = dispatch => ({
    addItem: item => dispatch(addItem(item))
});

We are creating this new function that is a prop called addItem that will go into our CollectionItem as the addItem function that we need to leverage.

addItem: item => dispatch(addItem(item)):
And then whenver we dispatch or call this function, this function will receive the 'item' as the property, pass it into our addItem action creator whihch gives us back that object where type is equal to addItem and the payload is equal to the item that we passed in and then we dispatch that new object into our store and it will go through our redux flow.