We are gonna look inside of our existing cart items to see if that cart item already exists: const existingCartItem = cartItems.find();
cartItems.find() will return you the first item found in our array based off a condition that we pass in find(here).

we get each individual item (cartItem) and we check the cartItem id if it matches the cartItemToAdd id:
cartItems.find(cartItem => cartItem.id === cartItemToAdd.id);

If it matches, it will set that cartItem where this condition (cartItem.id === cartItemToAdd.id) is true to our constant 'existingCartItem'. If it doesn't find anything, it will be undefined.