we check if that cart item already exists: if (existingCartItem)
cartItems.map() will return us a new array and we need to return new versions of our state
so that our components know to re-render properly.

we check if the we already have that cartItem: cartItem.id === cartItemToAdd.id?

true: {...cartItem, quantity: cartItem.quantity + 1}
we increase the amount (we don't have a quantity yet but we'll solve that later).

false: cartItem
we just wanna return the original item because there's no need to update any of the componentsthat might be relying on our object.