If the cartItem is not found inside of our array then:
we return a new array with all of the existing items that were already thereand also add an object  which is the item we wanna add (cartItemToAdd) with a based quantity of 1:
return [...cartItems, {...cartItemToAdd, quantity: 1}];

This way, whenever we add our cartItems in, any subsequent cartItems we'll be able to reference that quantity value that we have added to our objects when we store them.

(quantity property gets attached the first time around since this if block will not run when it's a new item).