now we can destructure itemCount off of our props and add it to the CartItcon component.
what we've written is called a "Selector":

//we are increasibng the number of the cartIcon everytime a new item is added:
const mapStateToProps = ({ cart: { cartItems } }) => ({
    itemCount: cartItems.reduce((accumulatedQuanttity, cartItem) => accumulatedQuanttity + cartItem.quantity, 0)
});

Because what we're doing is we are writing some code that gets a state as in the whole state object and then pulls off just a small portion or slice of that state because we are getting the 'cart' and then the 'cartItems' and then reducing all those cartItems that get a new value.

We are computing out a new value based off of the state.

The problem is that whenever any reducer updates, whenever we retunr a new object and redux recomposes and rebuilds the entire state object, mapStateToProps is getting called every single time, whihch is always passing in new props to our components. This is always rerendering components.