installed reselect.
Selector: code that gets a state and just pulls off a small portion of that state.

We don't wanna re render our components everytime the state changes, especially if those state changes don't actually modify the parts of the state that our component cares about. So we wanna be able to store or catch the value  of what our selector is using to compute its value.
For example: if the cartItem values doesn't change or the output of the selector doesn't change, we don'r wanna rerender our component. This is called Memorization.

Memorization is the catching of the selector's value.
We can achieve this by using a library called 'Reselect'.

'Reselect' allows us to write these selectors in such a way so that it knows that if the properties that it's pulling from the state and using are the same (their value doesn't change and the output of the selector is not different) then it will not pass them into the component. it will just pass the old value and our React component will not rerender.

It also allows us to separate our selectors and make them reusable.