created cart.selectors.js file inside of redux/cart
we have to switch up our folder structure because we are gonna move the code that we normally write inside mapStateToProps into their own folder inside of the respective redux folder that this will be related to.

For example: 'itemCount' can be a reusable selector, because we might want itemCount in multiple places of our application.