imported withRouter in cart-dropdown and wrapped: withRouter(connect(mapStateToProps)(CartDropdown))
The reason we can do this is because all of our higher order components return components but also take components as their arguments.

So withRouter() is just taking the component that got returned from our connect call as its component argument.

The order in which we  wrap them matters because withRouter() will be what passes the match history and location objects into the component that is being wrapped (connect(mapStateToProps)(CartDropdown))

So we want what comec out of the connect component to receive those props because when you wrap components like this, it actually evaluates from inside out.

So we get our connected component first which will then pass in to withRouter() higher order component.

And this way our actual component will have access to the props that we want which is: "history".

And we can use this function:
() => history.push('/checkout')
to take us to '/checkout' (CheckoutPage).