we wanna trigger a toggle whenever we click our 'go to checkout' button to toggle the hidden state of our cart.
So we import our togglecartHidden action in our cart-dropdown.

Now we would write mapDispatchToProps() and then pass in a dispatched call of our action creator of toggleCartHidden.

What we might not know is that 'connect' passes dispatch into our components as a prop if we do not supply a 2 argument to our 'connect()'. So if we don't supply mapDispatchToProps as the second parameter, 'connect' will pass dispatch into our 'CartDropdown' component as a property.

So we can do this:

const CartDropdown = ({ cartItems, history, dispatch })

<CustomButton onClick={() => {
            history.push('/checkout');
            dispatch(toggleCartHidden());}}
        >
                GO TO CHECKOUT
        </CustomButton>

Now when we click on 'go to checkout', our toggleCartHidden() is being fired.