Our remove-button is gonna be a UTF 8 winging.
UTF 8 is a standard that most browsers have that represents what charecters they are able to display or they're able to recognize.

So all the letters that we type that we're most familiar with, those all fall into the UTF 8 categorization.

But there are also additional symbols: https://www.w3schools.com/charsets/ref_utf_dingbats.asp

that we can that our browser knows what they represent.

So we do:
&#10054;

This will display a symbol that we got from that list, you just have to type the number and: &# number ;