added new case in cartReducer (clear item from cart) where we return a new array without any instance of that item that we're trying to clear away.
The code that we need to write in our reducer is a 'filter':

state.cartItems.filter(
                    cartItem => cartItem.id !== action.payload.id
                    )

If the cartItem.id doesn't match the item we are trying to remove (action.payload.id) then i want you to keep it.
If the cartItem does match the item's id that we're trying to remove from our payload, then i want you to filter it out.

filter() returns us back anything that yields TRUE inside of the actual function as the return of the function.

Basically we are saying:
cartItem => cartItem.id !== action.payload.id

If the cartItem doesn't match the item that we try to remove, then return true, so we'll keep that item.