created REMOVE_ITEM action type.
created removeItem() action.

added REMOVE_ITEM case in our cart reducer.

created removeItemFromCart() function where we grab the existingCartItem that we have in our cart by using find() and comparing id and then we check that if that existingCartItem we have only 1 if we remove it we want to filter all the rest of the items except for the one that we removed.

otherwise we will return all the cartItems, including the one that we removed but we will decrease the quantity of that item subtracting (-) 1.