installed redux-persist for localStorage and sessionStorage.
yarn add redux-persist

(redux-persist is used for localStorage [when we leave the website and we go back, the things that we have done are still there,
like when we add items to a cart] and sessionStorage [when we refresh the page, the things that we have done are still there
unless we close session ]).

The read-only localStorage property allows you to access a Storage object for the Document's origin;
the stored data is saved across browser sessions. localStorage is similar to sessionStorage, except that
while data stored in localStorage has no expiration time, data stored in sessionStorage gets cleared when
the page session ends — that is, when the page is closed. (Data in a localStorage object created in a
"private browsing" or "incognito" session is cleared when the last "private" tab is closed.)