imported persistStore.
persistStore allows our browser to cache our store depending on certain configuration options that we're gonna set.

First inside of the store.js we're bringing in persistStore.

We export the store and then:
export const persistor = persistStore(store);

So 'persistor' is a persisted version of our store.

And by doing this we create our new provider that's wrapping our application.

And we export an object with both store and persistor.

now we have access to both of these if we need it.

Also we removed the 'export' from const store since we don't need because we are already exporting here:

export default { store, persistor };