removed somethings from ShopPage and pass the CollectionsOverview component.
Also fixed some issues in collections-overview.

Everything works as before, except now our collection-preview it's in its own component.

So now for our shop.component we are gonna selectively render inside of nested routes where one route is gonna lead us to our collections-overview when we're only on shop.

So nothing but just /shop but then we'll also render our new category page which we'll build when we land on our /category:

/shop/:category