created collection component.
imported CollectionPage in the shop component.

Added new Route to CollectionPage where the path will be: match.path(/shop) and then we add '/:collectionId' where we will pass an id of each collection so that we go to that spefic page (hats page) depending on the id of each collection.