created COLLECTION_ID_MAP which is just an object that maps the string value to the respective ID where the string that we are getting from our URL parameter (/shop/hats) will be the actual property.
COLLECTION_ID_MAP = {
    gloves: 1,
    pants: 2,
    jackets: 3,
    boots: 4,
    snowboards: 5
}

So we are saying:
gloves goes to the id of 1,
pants goes to the id of 2,
etc.

So we are gonna pass in that string and then use that as the dynamic value of the property to get the correct ID and then  we'll match it in our selector.

So let's write that selector...