Changed width of collection-item in styles.
destructured  title and items off of our collection.

added title and mapped through the items and passing them to our CollectionItem component so that we display them once we click on hats, gloves or other different collection.

Remove the memoize() function from shop.selectors.js since it didn't work that way and we go back to the previous way with COLLECTION_ID_MAP.