DATA NORMALIZATION
turned shop_data into an object (instead of an array).

removed COLLECTION_ID_MAP from shop selectors.

And changed the selectCollection selector by not using the find() function and directly accesing the collections like this:
collections => collections[collectionUrlParam]

We are passing the url parameter into our selector
and our selector goes into this collections object and gets the correcponding value based on the properties gloves, hats, etc.

This is the ideal way that we should be storing large arrays of elements where we need the individual element inside of our reducer.

We store lists of elements as objects instead of arrays.