created selectCollectionsForPreview() selector to turn our collections object into an array and then map through each collection (hats, jackets...).
So we are saying:
Object.keys(collections).map(key => collections[key])

We wanna get all the keys and then map over that array of keys so we get the value of our collections object at that key which will give us an array of our items that we're trying to get.

This is pretty much the same way that we had it when we're trying to call our original selectCollections when it was an array.