collectionsArray.map(({title, items}) => ({ title, items }))
we pass a new array where we get the object and we destrcuture off of it the title and the items and we return a new objetc where title goes to title and items goes to items.

So we are passing only the values that we wanna keep.

And this way we have saved our documents (items) inside of our collections inside of our firebase database.