the snapshot is the snapshot representation of our collections array that we are gonna get from firestore. And we are gonna fetch taht, inside of our componentDidMount().
Imported firestore library from our firestore utils.

Now that we have this, inside of our componentDidMount() we do:

const collectionRef = firestore.collection('collections');
'collection' is the title of our collections inside of firestore.

And now we wanna get this data by using the same method that we use for our userRef which is onSnapshot():

collectionRef.onSnapshot(async snapshot => {console.log(snapshot);});

collectionRef.onSnapshot: whenever thos code gets run for the first time, collectionRef will send us the snapshot representing the code of our collections object array at the time when this code renders.
But we wanna perform an asynchronous request because the data is inside the actual object inside of this query snapshot, so we do async.