convertCollectionsSnapshotToMap = (collections):
This function gets the whole snapshot object of our collections.

we wanna convert it to an object instead of the array that we get.

This transforms the doc objects (array of wuery snapshot documents) and we'll return the right object with the right properties and our new route name property.

 const transformedCollection = collections.docs.map()

docs: gives us our query snapshot array

map(): and we map through this docs

doc => {
      const {title, items } = doc.data();

      We pull off the title and the items from our doc.data()

And we return an object that represents the final object representing all the data we want for our front end:

return {
        routeName: encodeURI(title.toLowerCase()),
        id: doc.id,
        title,
        items
      }

encodeURI(): you pass some string and it will give you back where any characters that a URL cannot handle or process such as certain symbols, it will make sure to convert them into sth that a URL can read.