return transformedCollection.reduce((accumulator, collection) => {       accumulator[collection.title.toLowerCase()] = collection;       return accumulator;     }, {})
  {} = we pass in an initial object. The initial object goes into the first new collection (first element)

accumulator[collection.title.toLowerCase()]: sets the first value equal to the title, so the first value is gonna be hats, that is equal to the hats collection.

return accumulator: and the  we return that object.
the title of the five objects are the keys  and then they equal their respective collection object.

we console log that object that we get in our shop component (where we are using it).