we wanna store the collections object in our shop reducer because now…
… that we have our modification we just wanna pass this and update our reducer.

we are gonna update our reducer and redux code to now be able to fire actions that update the reducer.

Now we have our new collections with the ids from our backend with the shape and format that we are looking for

We've transformed our array into the map and updated our shop redux files to be able to update and store this new collections map object.