added styles for our spinner and ccreated our spinner component.
This WithSpinner is a Higher Order Component.

This way we pass through the props to the component that we wrap:
<WrappedComponent {...otherProps} />

We are making a new withspinner component, a higher order component. It's a function that takes some component (wrappedcomponent) that we wanna wrap with the functionality of our spinner loading feature.

That wrapped component gets passed into this new component  that wraps around it.

It turns based on this isLoading property that gets passed into this component...

if isLoading  render my spinneroberlay:
isLoading ? (
        <SpinnerOverlay>

if isLoading is false, then just render the component that we passed in as normal receiving all the props that it would receive (otherProps)