imported withspinner component in shop component.
added state loading: true, in shop component.

added: this.setState({ loading: false});, iniside our componentDidMount().

WithSpinner(CollectionsOverview): This will give us back a new component wrapped around the component that we passed in.

And using this new component, we can just pass the loading value.

We do the same thing with our collectionpage:

const CollectionPageWithSpinner = WithSpinner(CollectionPage);

we destructure off the loading value from our state:
 const { loading } = this.state;

we use render value instead of the component value.

'render' is a method that takes a function where the parameters are the ones that the component will receive which in our case will be the 'match', 'location' and 'history' props that our route passes into our component which gives our collection page access to the match object that we use inside of our selector.

- render={(props) => <CollectionsOverviewWithSpinner isLoading={loading} {...props} />} />

- render={(props) => <CollectionPageWithSpinner isLoading={loading} {...props} />} />