PROMISE PATTERN
changed from observer pattern to promise pattern.

We're just going to treat it like some regular API endpoint that we can make requests to or make API

calls to.

Because firebase allows us to do so there's actually two ways we can do this.

The first way is we still leverage the fire store collection ref objects that we get back from the fire

store or library except instead of returning this observable style of object we return promises.

I'm going to show you that way and then I'll show you the other way where we use the native fetch in

order for us to still use our collection rough to do so except doing it in a promise oriented manner

instead of using on snapshot what we use is dot get what dot get does for us is it makes a API call

to fetch back the data associated to this collection.

Which will it be the exact same as the snapshot object that we're getting back from are back in.

The only difference is that instead now we call dot then and we pass our Snapshot transformation into this .then()